moved logging to utility class, removed analyticsjsonkey stringdef because it wasn't used and didn't make sense since we expect custom keys too
